# internal/http/handlers/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: HTTP API for the Codevertex library management service. Catalog, circulation,
    members, fines and e-books.
  title: Library Service API
  version: 0.1.0
paths:
...
  /{tenant}/library/catalog/sru/search:
    get:
      parameters:
      - description: search terms (title/author/ISBN)
        in: query
        name: q
        required: true
        type: string
      - description: SRU base URL (defaults to LoC)
        in: query
        name: target
        type: string
      responses: {}
      summary: 'Search an external SRU catalog (default: Library of Congress) for
        copy-cataloging'
      tags:
      - Catalog
...
schemes:
- http
- https
securityDefinitions:
  bearerAuth:
    description: 'JWT token from auth-service. Format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"